Port unicode to temp mode
